Compatibility with Redmine 5.0
