+++ app/models/kb_mailer.rb
@@ -1,6 +1,6 @@
 class KbMailer < Mailer
 
-  add_template_helper(KnowledgebaseHelper)
+  respond_to?(:add_template_helper) ? add_template_helper(KnowledgebaseHelper) : helper(KnowledgebaseHelper)
 
   def article_create(_user = User.current, article)
 	redmine_headers 'Project' => article.project.identifier
