+++ db/migrate/20100317201659_add_ratings_to_articles.rb
@@ -1,6 +1,4 @@
 class AddRatingsToArticles < Rails.version < '5.1' ? ActiveRecord::Migration : ActiveRecord::Migration[4.2]
-  require 'acts_as_rated'
-
   def self.up
     ActiveRecord::Base.create_ratings_table :with_rater => false
   end
