+++ db/migrate/20150326093122_add_taggings_counter_cache_to_tags.rb
@@ -9,6 +9,5 @@ class AddTaggingsCounterCacheToTags < Rails.version < '5.1' ? ActiveRecord::Migr
   end
 
   def self.down
-    remove_column :tags, :taggings_count
   end
 end
