+++ lib/redmine_knowledgebase/concerns/knowledgebase_project_extension.rb
@@ -1,0 +1,11 @@
+module RedmineKnowledgebase
+  module Concerns
+    module KnowledgebaseProjectExtension
+      extend ActiveSupport::Concern
+      included do
+        has_many :categories, :class_name => "KbCategory", :foreign_key => "project_id"
+        has_many :articles, :class_name => "KbArticle", :foreign_key => "project_id"
+      end
+    end
+  end
+end
